feat(oxc): remove `oxc_napi` crate
napi has a limitation, it cannot feature gate type exposure.
