+++ .github/workflows/ci.yml
@@ -285,11 +285,12 @@ jobs:
       - uses: Boshen/setup-rust@main
         if: steps.filter.outputs.src == 'true'
         with:
-          save-cache: ${{ github.ref_name == 'main' }}
           cache-key: warm
       - uses: ./.github/actions/pnpm
         if: steps.filter.outputs.src == 'true'
-      - run: pnpm run build
+      - run: pnpm run build-dev
         if: steps.filter.outputs.src == 'true'
       - run: pnpm run test
+        if: steps.filter.outputs.src == 'true'
+      - run: git diff --exit-code # Must commit everything
         if: steps.filter.outputs.src == 'true'
