+++ Cargo.toml
@@ -91,7 +91,6 @@ oxc_estree = { version = "0.38.0", path = "crates/oxc_estree" }
 oxc_isolated_declarations = { version = "0.38.0", path = "crates/oxc_isolated_declarations" }
 oxc_mangler = { version = "0.38.0", path = "crates/oxc_mangler" }
 oxc_minifier = { version = "0.38.0", path = "crates/oxc_minifier" }
-oxc_napi = { version = "0.38.0", path = "crates/oxc_napi" }
 oxc_parser = { version = "0.38.0", path = "crates/oxc_parser" }
 oxc_regular_expression = { version = "0.38.0", path = "crates/oxc_regular_expression" }
 oxc_semantic = { version = "0.38.0", path = "crates/oxc_semantic" }
@@ -109,9 +108,9 @@ oxc_tasks_common = { path = "tasks/common" }
 oxc_tasks_transform_checker = { path = "tasks/transform_checker" }
 
 # Relaxed version so the user can decide which version to use.
-napi = "3.0.0-alpha"
+napi = "3.0.0-alpha.11"
 napi-build = "2.1.3"
-napi-derive = "3.0.0-alpha"
+napi-derive = "3.0.0-alpha.11"
 
 # Relaxed version so the user can decide which version to use.
 proc-macro2 = "1"
