--- crates/oxc_napi/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "oxc_napi"
-version = "0.38.0"
-authors.workspace = true
-categories.workspace = true
-edition.workspace = true
-homepage.workspace = true
-include = ["/src"]
-keywords.workspace = true
-license.workspace = true
-publish = false
-repository.workspace = true
-rust-version.workspace = true
-description.workspace = true
-
-[lints]
-workspace = true
-
-[lib]
-test = false
-doctest = false
-
-[dependencies]
-oxc_isolated_declarations = { workspace = true }
-oxc_sourcemap = { workspace = true, features = ["napi"] }
-oxc_span = { workspace = true }
-oxc_syntax = { workspace = true }
-oxc_transformer = { workspace = true }
-
-napi = { workspace = true }
-napi-derive = { workspace = true }
-
-rustc-hash = { workspace = true }
