--- crates/oxc_napi/src/lib.rs
@@ -1,5 +1,0 @@
-pub mod parse;
-
-pub mod isolated_declarations;
-
-pub mod transform;
