+++ napi/minify/package.json
@@ -2,6 +2,7 @@
   "name": "@oxc-minify/binding",
   "private": true,
   "scripts": {
+    "build-dev": "napi build --platform",
     "build": "napi build --platform --release",
     "test": "echo 'skip'"
   },
