+++ napi/parser/Cargo.toml
@@ -21,12 +21,13 @@ test = false
 doctest = false
 
 [dependencies]
-oxc = { workspace = true, features = ["napi", "serialize"] }
-oxc_napi = { workspace = true }
+oxc = { workspace = true, features = ["serialize"] }
+
+rustc-hash = { workspace = true }
+serde_json = { workspace = true }
 
 napi = { workspace = true, features = ["async"] }
 napi-derive = { workspace = true }
-serde_json = { workspace = true }
 
 [package.metadata.cargo-shear]
 ignored = ["napi"]
