+++ napi/parser/index.js
@@ -1,7 +1,5 @@
 const bindings = require('./bindings.js');
 
-module.exports.moduleLexerAsync = bindings.moduleLexerAsync;
-module.exports.moduleLexerSync = bindings.moduleLexerSync;
 module.exports.parseWithoutReturn = bindings.parseWithoutReturn;
 
 module.exports.parseAsync = async function parseAsync(...args) {
