+++ napi/parser/src/convert.rs
@@ -1,208 +1,15 @@
-use napi_derive::napi;
-
 use rustc_hash::FxHashMap;
 
-use oxc_span::Span;
-use oxc_syntax::module_record::{self, ModuleRecord};
+use oxc::{
+    span::Span,
+    syntax::module_record::{self, ModuleRecord},
+};
 
-#[napi(object)]
-#[derive(Default)]
-pub struct ParserOptions {
-    #[napi(ts_type = "'script' | 'module' | 'unambiguous' | undefined")]
-    pub source_type: Option<String>,
-
-    /// Treat the source text as `js`, `jsx`, `ts`, or `tsx`.
-    #[napi(ts_type = "'js' | 'jsx' | 'ts' | 'tsx'")]
-    pub lang: Option<String>,
-
-    /// Emit `ParenthesizedExpression` in AST.
-    ///
-    /// If this option is true, parenthesized expressions are represented by
-    /// (non-standard) `ParenthesizedExpression` nodes that have a single `expression` property
-    /// containing the expression inside parentheses.
-    ///
-    /// Default: true
-    pub preserve_parens: Option<bool>,
-}
-
-#[napi(object)]
-#[derive(Default)]
-pub struct ParseResult {
-    #[napi(ts_type = "import(\"@oxc-project/types\").Program")]
-    pub program: String,
-    pub module: EcmaScriptModule,
-    pub comments: Vec<Comment>,
-    pub errors: Vec<String>,
-}
-
-#[napi(object)]
-pub struct Comment {
-    #[napi(ts_type = "'Line' | 'Block'")]
-    pub r#type: &'static str,
-    pub value: String,
-    pub start: u32,
-    pub end: u32,
-}
-
-#[napi(object)]
-#[derive(Default)]
-pub struct EcmaScriptModule {
-    /// Import Statements.
-    pub static_imports: Vec<StaticImport>,
-    /// Export Statements.
-    pub static_exports: Vec<StaticExport>,
-}
-
-#[napi(object)]
-pub struct ValueSpan {
-    pub value: String,
-    pub start: u32,
-    pub end: u32,
-}
-
-#[napi(object)]
-pub struct StaticImport {
-    /// Start of import statement.
-    pub start: u32,
-    /// End of import statement.
-    pub end: u32,
-    /// Import source.
-    ///
-    /// ```js
-    /// import { foo } from "mod";
-    /// //                   ^^^
-    /// ```
-    pub module_request: ValueSpan,
-    /// Import specifiers.
-    ///
-    /// Empty for `import "mod"`.
-    pub entries: Vec<ImportEntry>,
-}
-
-#[napi(object)]
-pub struct ImportEntry {
-    /// The name under which the desired binding is exported by the module.
-    ///
-    /// ```js
-    /// import { foo } from "mod";
-    /// //       ^^^
-    /// import { foo as bar } from "mod";
-    /// //       ^^^
-    /// ```
-    pub import_name: ImportName,
-    /// The name that is used to locally access the imported value from within the importing module.
-    /// ```js
-    /// import { foo } from "mod";
-    /// //       ^^^
-    /// import { foo as bar } from "mod";
-    /// //              ^^^
-    /// ```
-    pub local_name: ValueSpan,
-
-    /// Whether this binding is for a TypeScript type-only import.
-    ///
-    /// `true` for the following imports:
-    /// ```ts
-    /// import type { foo } from "mod";
-    /// import { type foo } from "mod";
-    /// ```
-    pub is_type: bool,
-}
-
-#[napi(string_enum)]
-pub enum ImportNameKind {
-    /// `import { x } from "mod"`
-    Name,
-    /// `import * as ns from "mod"`
-    NamespaceObject,
-    /// `import defaultExport from "mod"`
-    Default,
-}
-
-#[napi(object)]
-pub struct ImportName {
-    pub kind: ImportNameKind,
-    pub name: Option<String>,
-    pub start: Option<u32>,
-    pub end: Option<u32>,
-}
-
-#[napi(object)]
-pub struct StaticExport {
-    pub start: u32,
-    pub end: u32,
-    pub entries: Vec<ExportEntry>,
-}
-
-#[napi(object)]
-pub struct ExportEntry {
-    pub start: u32,
-    pub end: u32,
-    pub module_request: Option<ValueSpan>,
-    /// The name under which the desired binding is exported by the module`.
-    pub import_name: ExportImportName,
-    /// The name used to export this binding by this module.
-    pub export_name: ExportExportName,
-    /// The name that is used to locally access the exported value from within the importing module.
-    pub local_name: ExportLocalName,
-}
-
-#[napi(string_enum)]
-pub enum ExportImportNameKind {
-    /// `export { name }
-    Name,
-    /// `export * as ns from "mod"`
-    All,
-    /// `export * from "mod"`
-    AllButDefault,
-    /// Does not have a specifier.
-    None,
-}
-
-#[napi(object)]
-pub struct ExportImportName {
-    pub kind: ExportImportNameKind,
-    pub name: Option<String>,
-    pub start: Option<u32>,
-    pub end: Option<u32>,
-}
-
-#[napi(string_enum)]
-pub enum ExportExportNameKind {
-    /// `export { name }
-    Name,
-    /// `export default expression`
-    Default,
-    /// `export * from "mod"
-    None,
-}
-
-#[napi(object)]
-pub struct ExportExportName {
-    pub kind: ExportExportNameKind,
-    pub name: Option<String>,
-    pub start: Option<u32>,
-    pub end: Option<u32>,
-}
-
-#[napi(object)]
-pub struct ExportLocalName {
-    pub kind: ExportLocalNameKind,
-    pub name: Option<String>,
-    pub start: Option<u32>,
-    pub end: Option<u32>,
-}
-
-#[napi(string_enum)]
-pub enum ExportLocalNameKind {
-    /// `export { name }
-    Name,
-    /// `export default expression`
-    Default,
-    /// If the exported value is not locally accessible from within the module.
-    /// `export default function () {}`
-    None,
-}
+use crate::types::{
+    EcmaScriptModule, ExportExportName, ExportExportNameKind, ExportImportName,
+    ExportImportNameKind, ExportLocalName, ExportLocalNameKind, ImportName, ImportNameKind,
+    StaticExport, StaticExportEntry, StaticImport, StaticImportEntry, ValueSpan,
+};
 
 impl From<&ModuleRecord<'_>> for EcmaScriptModule {
     fn from(record: &ModuleRecord<'_>) -> Self {
@@ -215,7 +22,7 @@ impl From<&ModuleRecord<'_>> for EcmaScriptModule {
                         .import_entries
                         .iter()
                         .filter(|e| e.statement_span == m.statement_span)
-                        .map(ImportEntry::from)
+                        .map(StaticImportEntry::from)
                         .collect::<Vec<_>>();
                     {
                         StaticImport {
@@ -239,10 +46,10 @@ impl From<&ModuleRecord<'_>> for EcmaScriptModule {
             .iter()
             .chain(record.indirect_export_entries.iter())
             .chain(record.star_export_entries.iter())
-            .map(|e| (e.statement_span, ExportEntry::from(e)))
+            .map(|e| (e.statement_span, StaticExportEntry::from(e)))
             .collect::<Vec<_>>()
             .into_iter()
-            .fold(FxHashMap::<Span, Vec<ExportEntry>>::default(), |mut acc, (span, e)| {
+            .fold(FxHashMap::<Span, Vec<StaticExportEntry>>::default(), |mut acc, (span, e)| {
                 acc.entry(span).or_default().push(e);
                 acc
             })
@@ -255,7 +62,20 @@ impl From<&ModuleRecord<'_>> for EcmaScriptModule {
     }
 }
 
-impl From<&module_record::ImportEntry<'_>> for ImportEntry {
+impl From<&module_record::ExportEntry<'_>> for StaticExportEntry {
+    fn from(e: &module_record::ExportEntry) -> Self {
+        Self {
+            start: e.span.start,
+            end: e.span.end,
+            module_request: e.module_request.as_ref().map(ValueSpan::from),
+            import_name: ExportImportName::from(&e.import_name),
+            export_name: ExportExportName::from(&e.export_name),
+            local_name: ExportLocalName::from(&e.local_name),
+        }
+    }
+}
+
+impl From<&module_record::ImportEntry<'_>> for StaticImportEntry {
     fn from(e: &module_record::ImportEntry<'_>) -> Self {
         Self {
             import_name: ImportName::from(&e.import_name),
@@ -291,19 +111,6 @@ impl From<&module_record::NameSpan<'_>> for ValueSpan {
             value: name_span.name.to_string(),
             start: name_span.span.start,
             end: name_span.span.end,
-        }
-    }
-}
-
-impl From<&module_record::ExportEntry<'_>> for ExportEntry {
-    fn from(e: &module_record::ExportEntry) -> Self {
-        Self {
-            start: e.span.start,
-            end: e.span.end,
-            module_request: e.module_request.as_ref().map(ValueSpan::from),
-            import_name: ExportImportName::from(&e.import_name),
-            export_name: ExportExportName::from(&e.export_name),
-            local_name: ExportLocalName::from(&e.local_name),
         }
     }
 }
