+++ napi/transform/src/lib.rs
@@ -1,7 +1,5 @@
 mod errors;
 
-pub use oxc_napi::{isolated_declarations, transform};
-
 mod isolated_declaration;
 pub use isolated_declaration::*;
