+++ package.json
@@ -7,6 +7,7 @@
   },
   "scripts": {
     "build": "pnpm --workspace-concurrency=1 --filter './napi/*' build",
+    "build-dev": "pnpm --workspace-concurrency=1 --filter './napi/*' build-dev",
     "test": "pnpm --workspace-concurrency=1 --filter './napi/*' test"
   },
   "devDependencies": {
